+++ CHANGELOG.md
@@ -9,6 +9,7 @@
       - "Number of dog sledge used for fishing (temporal)"
       - "Number of fishing vessels used for fishing (temporal)"
       - "Number of snow mobile used for fishing (temporal)"
+      - "Total fish and shellfish landings in tonnes (temporal)"
 
 # v4.0.0alpha3 (2025-03-26)